+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2021-04-27
+
+### Fixed
+
+- `EnsureNotProblemDetailAsync` should use configured casing for CorrelationId key name
+- `ReadAsProblemDetailsAsync` should fallback to a ProblemDetails rather than throwing
+
 ## [1.0.0] - 2021-04-27
 
 ### Added